% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgcfg_db}
\alias{pkgcfg_db}
\title{pkg-config snapshot from install time}
\usage{
pkgcfg_db()
}
\description{
Returns a data.frame with columns
`module`, `cflags`, `libs`, `static_libs`.
}
\keyword{internal}
